refactor: FSD Phase 3 — widgets 層へ移す (FRESTYLE-159)
FSD 移行(FRESTYLE-154)の Phase 3。純粋な移動 + import 追従で振る舞いは変えない。

Slice 構成:
- widgets/app-shell/      AppShell / Header / HeaderUserMenu / CommandPalette
                          / SkipLink / ScrollToTop
- widgets/secondary-panel/ SecondaryPanel
- widgets/auth-layout/     AuthLayout

各 Slice に index.ts(名前付き re-export)を追加。

境界 lint が実違反を検出したため設計を修正:
当初 header / command-palette を別 Slice にしたところ、app-shell から参照して
同一層の Slice 間 import(FSD 禁止)になった。実利用を確認すると両者とも
AppShell からしか使われていないため、1 つの Slice に統合した
(常に一緒に使われるものは同じ Slice に置く)。

死んだコードの削除:
components/layout/UserAvatar.tsx は自分のテストからしか import されておらず、
実際のヘッダーは別コンポーネント(components/Avatar)を使用していた。新構造へ
持ち込まないためテストごと削除(参照が無いため振る舞い中立)。テスト件数が
1355→1350 に減るのはこのぶん。

検証: tsc / eslint(境界違反 0) / vitest(165ファイル1350件・assert 変更なし) /
build / E2E(12件) すべて緑。
