+++ frontend/src/widgets/app-shell/index.ts
@@ -1,0 +1,3 @@
+export { default as AppShell } from './ui/AppShell';
+export { default as SkipLink } from './ui/SkipLink';
+export { default as ScrollToTop } from './ui/ScrollToTop';
