+++ frontend/src/widgets/auth-layout/__tests__/AuthLayout.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import AuthLayout from '../AuthLayout';
+import AuthLayout from '../ui/AuthLayout';
 
 describe('AuthLayout', () => {
   it('タイトルが表示される', () => {
