+++ frontend/src/widgets/secondary-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as SecondaryPanel } from './ui/SecondaryPanel';
