+++ frontend/src/widgets/secondary-panel/ui/SecondaryPanel.tsx
